+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## [4.1.0](https://github.com/plivo/plivo-java/tree/v4.1.0) (2018-08-21)
+- Added api's for address and identity verification
+
+## [4.0.4](https://github.com/plivo/plivo-java/tree/v4.0.4) (2018-08-15)
+- Fixed asynchronous outbound call
+
+## [4.0.3](https://github.com/plivo/plivo-java/tree/v4.0.3) (2018-07-27)
+- Add gradle task `fatJar` to create fat jar.
+- Fix gradle task `javadocJar`.
+
 ## [4.0.2](https://github.com/plivo/plivo-java/tree/v4.0.2) (2018-04-25)
 - Fix validate Signature. Supporting URL with Port.
 - Add .create(), .get(), .delete(), and other actions to Resource.client(client).